修正enum和其他方法冲突

+++ src/XUnitFile2Object/Test1/UnitTest1.cs
@@ -11,7 +11,7 @@ namespace XUnitFile2Object.Test1
     {
         public int Code { get; set; }
         public string Name { get; set; }
-        public int Achievement { get; set; }
+        public decimal Achievement { get; set; }
 
         public DateTime AchievementCreateTime { get; set; }
     }
